Split packages into folders
