+++ .gitignore
@@ -1,5 +1,5 @@
 *
-!.java
+!*.java
 !pom.xml
 
 !.gitignore
