+++ src/main/java/lab1/Main.java
@@ -1,5 +1,7 @@
 package lab1;
 
+import lab1.shell.Shell;
+
 public class Main {
     public static void main(String[] args) {
         Shell.run();
