+++ src/main/java/lab1/universityStructure/Faculty.java
@@ -1,4 +1,4 @@
-package lab1;
+package lab1.universityStructure;
 
 import java.util.ArrayList;
 import java.util.Optional;
