+++ src/main/java/lab1/universityStructure/StudyField.java
@@ -1,4 +1,4 @@
-package lab1;
+package lab1.universityStructure;
 
 public enum StudyField {
     MECHANICAL_ENGINEERING,
